Urlencode error messages sent to mr3/error page.
Use alternate form of Redirect that accepts query parameters as a Map to
ensure that error messages are properly url-encoded when redirecting to
mr3/error page. This addresses the "Invalid header value character"
error that could occur during signin as a result of an error messaging
containing an unencoded newline being added to the Location header
during the redirect.