Remove properties from user insert on login. Fixes #481
When a user logs in the insert method is called on UserDAL
with the OSM xml data. The OSM xml data does not include
any kind of properties so consequently the properties field
was always overwritten to empty in the DB. Since we only
need to store the OSM data with this method we do not need
to insert/update the properties field and can ignore it.