Remove hour wait for tasks in created status when fetching next
When fetching next tasks to work on, we only return tasks that
the user has not worked on in the last hour. If the task's
status is reset back to created this limit still applies. Do
not require an hour wait limit if the task is in a created
status so it can be worked on right away in this instance.