Add method to change task status based on search parameters
Add api PUT method `/tasks/changeStatus` that will change
task status on all tasks matching the search parameters.
Note: The only thing on the task affected by this change
is the task status (ie. who mapped it or how long it took,
etc. will not be touched). This method is for challenge
admins only.