Add API to fetch list of challenges for review
Add api '/review/challenges' which will fetch challenges that
have tasks that have reviews/review requests. From this list,
the front end will be able to narrow down the reviews to be
returned to the user by challenge (or even by project) when
fetching the full list of reviews.

Example use:
/review/challenges?reviewTasksType=1&tStatus=1&exclueOtherReviewers=true