set play idle timeout in conf to 1 hour
When doing an overpass query we let the user override the timeout
in their query. This timoeout only overrides the request timeout
and will not affect the idle timeout. By changing the idle timeout
to an hour in the conf, it will allow longer overpass queries to
run with overriden timeouts but since the default request timeout
in play is 2 minutes it will still enforce that restriction
normally.