Improve performance of next review task query
When fetching next task to review, a task id can provided
so that the task following (based on criteria) will be returned.
The query to determine the row number was pulling back the full
set of results when only the actual task needed to be returned.
Fixed query by wrapping in another select filtering by task id.