Create the CaffeineCache implementation
This is the implementation of a `CaffeineCache` (caffeine is a highly performant in-memory cache) that can be used based on the conf. The `BasicCache` was shown to be a limiter in several profiles and lacks a priority queue for item expiry.

Use the more performant cache by setting `maproulette.caching.type = "caffeine"` in the conf file. At this time the default will remain the BasicCache until this gets more testing.