Generate routes file dynamically based on file modification
This commit introduces a dynamic generation mechanism for the routes file in the sbt project. The routes file is now generated only when necessary, i.e., if any of the dependent files have been modified.

The `generateRoutesFile` task has been updated to track the last modified times of the dependent files and the generated routes file. If the routes file doesn't exist or any of the dependent files have been modified since the last build, the routes file is regenerated. Otherwise, if none of the dependent files have changed, the task reports that the Swagger routes file is up to date.