Enhance Leaderboard Availability During Refresh (#1101)
* Enhance Leaderboard Availability During Refresh

The initial approach to refreshing the leaderboard, which involved deleting all entries at once, led to data unavailability for users during updates. We've revised the process to sequentially delete and rebuild entries for each time duration. This adjustment helps to ensure that leaderboard data remains accessible to users while updates for other durations are in progress.

* Do the update to user_leaderboard and user_top_challenges at the same time

* Sql null comparison needs to be 'is null' and not '= null'